[ADD] moving the API views into a views sub-package within the API application to ease view navigation
