+++ src/interfaces/api/views/__init__.py
@@ -1,0 +1,30 @@
+from .datasets import (OrganismBindingSites,
+                       OrganismsBindingSites,
+                       RegulatorBindingSites,
+                       RegulatorsBindingSites,
+                       TRN,
+                       TRNList)
+from .documentation import best_practices
+from .effector import (EffectorDetail,
+                       EffectorList)
+from .evidence import (EvidenceDetail,
+                       EvidenceList)
+from .gene import (GeneDetail,
+                   GeneList)
+from .index import IndexView
+from .operon import (OperonDetail,
+                     OperonList)
+from .organism import (OrganismDetail,
+                       OrganismList)
+from .pathway import (PathwayDetail,
+                      PathwayList)
+from .publication import (PublicationDetail,
+                          PublicationList)
+from .regulator import (RegulatorDetail,
+                        RegulatorList)
+from .regulatory_family import (RegulatoryFamilyDetail,
+                                RegulatoryFamilyList)
+from .regulatory_interaction import (InteractionDetail,
+                                     InteractionsList)
+from .tfbs import (BindingSiteDetail,
+                   BindingSitesList)
