+++ src/interfaces/api/views/documentation.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def best_practices(request):
+    return render(request, 'api/best-practices.html')
